Major Refactoring changes & started multi Lang sup
