+++ config/bot_info/bot_language.py
@@ -1,0 +1,5 @@
+## TYPE YOUR LANGUAGE INITIALS OF CHOICE HERE ##
+
+CHOSEN_LANGUAGE = "EN"  # take the languages list bellow for refference
+
+################################################
