+++ config/bot_info/bot_name.py
@@ -1,0 +1,5 @@
+######   Type the name your bot will call itself here   ######
+
+BOT_NAME = "Mashiro"
+
+##############################################################
