+++ config/bot_info/bot_prefix.py
@@ -1,0 +1,5 @@
+######   Type the prefix that will be used to activate standard_commands here   ######
+
+BOT_PREFIX = "mr! "
+
+######################################################################################
