+++ config/server/new_server_member_standard_roles.py
@@ -1,0 +1,8 @@
+## Add, in order, the roles ID's of the roles that you want the bot to give when a user joins your Discord server bellow ##
+
+roles_ids = [
+    967849474405838848,
+    967693392433807420,
+    967677263430713344,
+    928409990589448212
+]
