+++ language/EN/language_EN.py
@@ -1,0 +1,53 @@
+from config.bot_info.bot_name import BOT_NAME
+
+bot_display_name = BOT_NAME
+
+help_command = {
+    # \/ Message #
+    f"Get help related to {bot_display_name}'s commands":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        f"Get help related to {bot_display_name}'s commands",
+    
+    # \/ Message #
+    "category":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "category",
+        
+    # \/ Message #
+    "Choose the subject you want to get help with":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Choose the subject you want to get help with",
+        
+    # \/ List #
+    "help_command_choice_list":
+        
+        # \/ Template
+        #["Music", "Dice Rolling"]  <<<<<<<<
+        
+        # \/ Following the Template, type the list in your language between the quotes bellow #
+        ["Music", "Dice Rolling"],
+        
+    # \/ Message #
+    "Something went wrong, if the problem persists please contact a Guild Master":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Something went wrong, if the problem persists please contact a Guild Master",
+}
+
+
+
+stream_alerts = {
+    # \/ Message #
+    "Streaming":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Streaming",
+        
+    # \/ Message #
+    "Watch the stream here":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Watch the stream here",
+        
+    # \/ Message #
+    "This streamer has classified their content as recommended for a mature audience":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "This streamer has classified their content as recommended for a mature audience",
+}
