+++ language/PTBR/language_PTBR.py
@@ -1,0 +1,53 @@
+from config.bot_info.bot_name import BOT_NAME
+
+bot_display_name = BOT_NAME
+
+help_command = {
+    # \/ Message #
+    f"Get help related to {bot_display_name}'s commands":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        f"Obtenha ajuda relacionada aos comandos de {bot_display_name}",
+    
+    # \/ Message #
+    "category":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "categoria",
+        
+    # \/ Message #
+    "Choose the subject you want to get help with":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Escolha o assunto que quer receber ajuda",
+        
+    # \/ List #
+    "help_command_choice_list":
+        
+        # \/ Template
+        #["Music", "Dice Rolling"]  <<<<<<<<
+        
+        # \/ Following the Template, type the list in your language between the quotes bellow #
+        ["Músicas", "Rolagem de Dados"],
+        
+    # \/ Message #
+    "Something went wrong, if the problem persists please contact a Guild Master":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Algo deu errado, se o problema persistir, favor contatar um Mestre da Guilda",
+}
+
+
+
+stream_alerts = {
+    # \/ Message #
+    "Streaming":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Transmitindo",
+        
+    # \/ Message #
+    "Watch the stream here":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Assista à stream aqui",
+        
+    # \/ Message #
+    "This streamer has classified their content as recommended for a mature audience":
+        # \/ Following the example, type the Message in your language between the quotes bellow #
+        "Este streamer classificou o próprio conteúdo como não recomendado para menores de idade",
+}
