+++ language/languages.py
@@ -1,0 +1,8 @@
+from language.PTBR import language_PTBR
+from language.EN import language_EN
+
+
+languages = {
+    "PTBR":language_PTBR,
+    "EN":language_EN,
+}
