+++ manager.py
@@ -1,5 +1,5 @@
 from nextcord.ext import commands
-from alerts import stream_alerts
+from alerts.stream_alerts.stream_alerts import StreamAlert
 
 
 class Manager(commands.Cog):
@@ -11,8 +11,8 @@ class Manager(commands.Cog):
     @commands.Cog.listener()
     async def on_ready(self):
         print(f"-------------------------\nEstou online como {self.bot.user}\n-------------------------")
-        stream_alerts.StreamData.send_alert_message.start(self)
-        stream_alerts.StreamData.update_alert_message.start(self)
+        StreamAlert.send_alert_message.start(self)
+        StreamAlert.update_alert_message.start(self)
