+++ music_player/player_utils/player_panel_message_id.py
@@ -1,0 +1,3 @@
+panel_message_id = {
+    "id":""
+}
