+++ music_player/player_utils/player_queue_history.py
@@ -1,0 +1,1 @@
+queue_history = []
