+++ utils/bot_data_utils.py
@@ -1,0 +1,3 @@
+bot_data = {
+    "display_name":""
+}
